+++ src/werkzeug/datastructures.py
@@ -17,6 +17,7 @@ from itertools import repeat
 from . import exceptions
 from ._compat import BytesIO
 from ._compat import collections_abc
+from ._compat import fspath
 from ._compat import integer_types
 from ._compat import iteritems
 from ._compat import iterkeys
@@ -29,11 +30,6 @@ from ._compat import text_type
 from ._compat import to_native
 from ._internal import _missing
 from .filesystem import get_filesystem_encoding
-
-try:
-    import pathlib
-except ImportError:
-    pathlib = None
 
 
 def is_immutable(self):
@@ -2948,20 +2944,23 @@ class FileStorage(object):
 
         For secure file saving also have a look at :func:`secure_filename`.
 
-        :param dst: a filename or open file object the uploaded file
-                    is saved to.
-        :param buffer_size: the size of the buffer.  This works the same as
-                            the `length` parameter of
-                            :func:`shutil.copyfileobj`.
+        :param dst: a filename, :class:`os.PathLike`, or open file
+            object to write to.
+        :param buffer_size: Passed as the ``length`` parameter of
+            :func:`shutil.copyfileobj`.
+
+        .. versionchanged:: 1.0
+            Supports :mod:`pathlib`.
         """
         from shutil import copyfileobj
 
         close_dst = False
-        if pathlib is not None and isinstance(dst, pathlib.PurePath):
-            dst = str(dst)
+        dst = fspath(dst)
+
         if isinstance(dst, string_types):
             dst = open(dst, "wb")
             close_dst = True
+
         try:
             copyfileobj(self.stream, dst, buffer_size)
         finally: